requirements.txt
graphene==2.1.8
graphene-django==2.7.1
Django==3.0.1
django-cors-headers==3.2.0
Pillow==6.2.1
pytz==2019.3
django-filter==2.2.0